+++ malis/malis_cpp.h
@@ -5,8 +5,8 @@ void connected_components_cpp(const int nVert,
                const int nEdge, const int* node1, const int* node2, const int* edgeWeight,
                int* seg);
 
-void malis_loss_weights_cpp(const int nVert, const int* seg,
-               const int nEdge, const int* node1, const int* node2, const float* edgeWeight,
+void malis_loss_weights_cpp(const int nVert, const uint64_t* seg,
+               const int nEdge, const uint64_t* node1, const uint64_t* node2, const float* edgeWeight,
                const int pos,
                int* nPairPerEdge);
